Move some test to long test suite
