+++ server/jetstream_cluster_4_test.go
@@ -831,650 +831,6 @@ func TestJetStreamClusterConsumerPauseSurvivesRestart(t *testing.T) {
 	checkTimer(leader)
 }
 
-func TestJetStreamClusterStreamOrphanMsgsAndReplicasDrifting(t *testing.T) {
-	checkInterestStateT = 4 * time.Second
-	checkInterestStateJ = 1
-	defer func() {
-		checkInterestStateT = defaultCheckInterestStateT
-		checkInterestStateJ = defaultCheckInterestStateJ
-	}()
-
-	type testParams struct {
-		restartAny       bool
-		restartLeader    bool
-		rolloutRestart   bool
-		ldmRestart       bool
-		restarts         int
-		checkHealthz     bool
-		reconnectRoutes  bool
-		reconnectClients bool
-	}
-	test := func(t *testing.T, params *testParams, sc *nats.StreamConfig) {
-		conf := `
-		listen: 127.0.0.1:-1
-		server_name: %s
-		jetstream: {
-			store_dir: '%s',
-		}
-		cluster {
-			name: %s
-			listen: 127.0.0.1:%d
-			routes = [%s]
-		}
-		server_tags: ["test"]
-		system_account: sys
-		no_auth_user: js
-		accounts {
-			sys { users = [ { user: sys, pass: sys } ] }
-			js {
-				jetstream = enabled
-				users = [ { user: js, pass: js } ]
-		    }
-		}`
-		c := createJetStreamClusterWithTemplate(t, conf, sc.Name, 3)
-		defer c.shutdown()
-
-		// Update lame duck duration for all servers.
-		for _, s := range c.servers {
-			s.optsMu.Lock()
-			s.opts.LameDuckDuration = 5 * time.Second
-			s.opts.LameDuckGracePeriod = -5 * time.Second
-			s.optsMu.Unlock()
-		}
-
-		nc, js := jsClientConnect(t, c.randomServer())
-		defer nc.Close()
-
-		cnc, cjs := jsClientConnect(t, c.randomServer())
-		defer cnc.Close()
-
-		_, err := js.AddStream(sc)
-		require_NoError(t, err)
-
-		pctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-
-		// Start producers
-		var wg sync.WaitGroup
-
-		// First call is just to create the pull subscribers.
-		mp := nats.MaxAckPending(10000)
-		mw := nats.PullMaxWaiting(1000)
-		aw := nats.AckWait(5 * time.Second)
-
-		for i := 0; i < 10; i++ {
-			for _, partition := range []string{"EEEEE"} {
-				subject := fmt.Sprintf("MSGS.%s.*.H.100XY.*.*.WQ.00000000000%d", partition, i)
-				consumer := fmt.Sprintf("consumer:%s:%d", partition, i)
-				_, err := cjs.PullSubscribe(subject, consumer, mp, mw, aw)
-				require_NoError(t, err)
-			}
-		}
-
-		// Create a single consumer that does no activity.
-		// Make sure we still calculate low ack properly and cleanup etc.
-		_, err = cjs.PullSubscribe("MSGS.ZZ.>", "consumer:ZZ:0", mp, mw, aw)
-		require_NoError(t, err)
-
-		subjects := []string{
-			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000000",
-			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000001",
-			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000002",
-			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000003",
-			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000004",
-			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000005",
-			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000006",
-			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000007",
-			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000008",
-			"MSGS.EEEEE.P.H.100XY.1.100Z.WQ.000000000009",
-		}
-		payload := []byte(strings.Repeat("A", 1024))
-
-		for i := 0; i < 50; i++ {
-			wg.Add(1)
-			go func() {
-				pnc, pjs := jsClientConnect(t, c.randomServer())
-				defer pnc.Close()
-
-				for i := 1; i < 200_000; i++ {
-					select {
-					case <-pctx.Done():
-						wg.Done()
-						return
-					default:
-					}
-					for _, subject := range subjects {
-						// Send each message a few times.
-						msgID := nats.MsgId(nuid.Next())
-						pjs.PublishAsync(subject, payload, msgID)
-						pjs.Publish(subject, payload, msgID, nats.AckWait(250*time.Millisecond))
-						pjs.Publish(subject, payload, msgID, nats.AckWait(250*time.Millisecond))
-					}
-				}
-			}()
-		}
-
-		// Rogue publisher that sends the same msg ID everytime.
-		for i := 0; i < 10; i++ {
-			wg.Add(1)
-			go func() {
-				pnc, pjs := jsClientConnect(t, c.randomServer())
-				defer pnc.Close()
-
-				msgID := nats.MsgId("1234567890")
-				for i := 1; ; i++ {
-					select {
-					case <-pctx.Done():
-						wg.Done()
-						return
-					default:
-					}
-					for _, subject := range subjects {
-						// Send each message a few times.
-						pjs.PublishAsync(subject, payload, msgID, nats.RetryAttempts(0), nats.RetryWait(0))
-						pjs.Publish(subject, payload, msgID, nats.AckWait(1*time.Millisecond), nats.RetryAttempts(0), nats.RetryWait(0))
-						pjs.Publish(subject, payload, msgID, nats.AckWait(1*time.Millisecond), nats.RetryAttempts(0), nats.RetryWait(0))
-					}
-				}
-			}()
-		}
-
-		// Let enough messages into the stream then start consumers.
-		time.Sleep(15 * time.Second)
-
-		ctx, cancel := context.WithTimeout(context.Background(), 45*time.Second)
-		defer cancel()
-
-		for i := 0; i < 10; i++ {
-			subject := fmt.Sprintf("MSGS.EEEEE.*.H.100XY.*.*.WQ.00000000000%d", i)
-			consumer := fmt.Sprintf("consumer:EEEEE:%d", i)
-			for n := 0; n < 5; n++ {
-				cpnc, cpjs := jsClientConnect(t, c.randomServer())
-				defer cpnc.Close()
-
-				psub, err := cpjs.PullSubscribe(subject, consumer, mp, mw, aw)
-				require_NoError(t, err)
-
-				time.AfterFunc(15*time.Second, func() {
-					cpnc.Close()
-				})
-
-				wg.Add(1)
-				go func() {
-					tick := time.NewTicker(1 * time.Millisecond)
-					for {
-						if cpnc.IsClosed() {
-							wg.Done()
-							return
-						}
-						select {
-						case <-ctx.Done():
-							wg.Done()
-							return
-						case <-tick.C:
-							// Fetch 1 first, then if no errors Fetch 100.
-							msgs, err := psub.Fetch(1, nats.MaxWait(200*time.Millisecond))
-							if err != nil {
-								continue
-							}
-							for _, msg := range msgs {
-								msg.Ack()
-							}
-							msgs, err = psub.Fetch(100, nats.MaxWait(200*time.Millisecond))
-							if err != nil {
-								continue
-							}
-							for _, msg := range msgs {
-								msg.Ack()
-							}
-							msgs, err = psub.Fetch(1000, nats.MaxWait(200*time.Millisecond))
-							if err != nil {
-								continue
-							}
-							for _, msg := range msgs {
-								msg.Ack()
-							}
-						}
-					}
-				}()
-			}
-		}
-
-		for i := 0; i < 10; i++ {
-			subject := fmt.Sprintf("MSGS.EEEEE.*.H.100XY.*.*.WQ.00000000000%d", i)
-			consumer := fmt.Sprintf("consumer:EEEEE:%d", i)
-			for n := 0; n < 10; n++ {
-				cpnc, cpjs := jsClientConnect(t, c.randomServer())
-				defer cpnc.Close()
-
-				psub, err := cpjs.PullSubscribe(subject, consumer, mp, mw, aw)
-				if err != nil {
-					t.Logf("ERROR: %v", err)
-					continue
-				}
-
-				wg.Add(1)
-				go func() {
-					tick := time.NewTicker(1 * time.Millisecond)
-					for {
-						select {
-						case <-ctx.Done():
-							wg.Done()
-							return
-						case <-tick.C:
-							// Fetch 1 first, then if no errors Fetch 100.
-							msgs, err := psub.Fetch(1, nats.MaxWait(200*time.Millisecond))
-							if err != nil {
-								continue
-							}
-							for _, msg := range msgs {
-								msg.Ack()
-							}
-							msgs, err = psub.Fetch(100, nats.MaxWait(200*time.Millisecond))
-							if err != nil {
-								continue
-							}
-							for _, msg := range msgs {
-								msg.Ack()
-							}
-
-							msgs, err = psub.Fetch(1000, nats.MaxWait(200*time.Millisecond))
-							if err != nil {
-								continue
-							}
-							for _, msg := range msgs {
-								msg.Ack()
-							}
-						}
-					}
-				}()
-			}
-		}
-
-		// Periodically disconnect routes from one of the servers.
-		if params.reconnectRoutes {
-			wg.Add(1)
-			go func() {
-				for range time.NewTicker(10 * time.Second).C {
-					select {
-					case <-ctx.Done():
-						wg.Done()
-						return
-					default:
-					}
-
-					// Force disconnecting routes from one of the servers.
-					s := c.servers[rand.Intn(3)]
-					var routes []*client
-					t.Logf("Disconnecting routes from %v", s.Name())
-					s.mu.Lock()
-					for _, conns := range s.routes {
-						routes = append(routes, conns...)
-					}
-					s.mu.Unlock()
-					for _, r := range routes {
-						r.closeConnection(ClientClosed)
-					}
-				}
-			}()
-		}
-
-		// Periodically reconnect clients.
-		if params.reconnectClients {
-			reconnectClients := func(s *Server) {
-				for _, client := range s.clients {
-					client.closeConnection(Kicked)
-				}
-			}
-
-			wg.Add(1)
-			go func() {
-				for range time.NewTicker(10 * time.Second).C {
-					select {
-					case <-ctx.Done():
-						wg.Done()
-						return
-					default:
-					}
-					// Force reconnect clients from one of the servers.
-					s := c.servers[rand.Intn(len(c.servers))]
-					reconnectClients(s)
-				}
-			}()
-		}
-
-		// Restarts
-		time.AfterFunc(10*time.Second, func() {
-			for i := 0; i < params.restarts; i++ {
-				switch {
-				case params.restartLeader:
-					// Find server leader of the stream and restart it.
-					s := c.streamLeader("js", sc.Name)
-					if params.ldmRestart {
-						s.lameDuckMode()
-					} else {
-						s.Shutdown()
-					}
-					s.WaitForShutdown()
-					c.restartServer(s)
-				case params.restartAny:
-					s := c.servers[rand.Intn(len(c.servers))]
-					if params.ldmRestart {
-						s.lameDuckMode()
-					} else {
-						s.Shutdown()
-					}
-					s.WaitForShutdown()
-					c.restartServer(s)
-				case params.rolloutRestart:
-					for _, s := range c.servers {
-						if params.ldmRestart {
-							s.lameDuckMode()
-						} else {
-							s.Shutdown()
-						}
-						s.WaitForShutdown()
-						c.restartServer(s)
-
-						if params.checkHealthz {
-							hctx, hcancel := context.WithTimeout(ctx, 15*time.Second)
-							defer hcancel()
-
-							for range time.NewTicker(2 * time.Second).C {
-								select {
-								case <-hctx.Done():
-								default:
-								}
-
-								status := s.healthz(nil)
-								if status.StatusCode == 200 {
-									break
-								}
-							}
-						}
-					}
-				}
-				c.waitOnClusterReady()
-			}
-		})
-
-		// Wait until context is done then check state.
-		<-ctx.Done()
-
-		getStreamDetails := func(t *testing.T, srv *Server) *StreamDetail {
-			t.Helper()
-			jsz, err := srv.Jsz(&JSzOptions{Accounts: true, Streams: true, Consumer: true})
-			require_NoError(t, err)
-			if len(jsz.AccountDetails) > 0 && len(jsz.AccountDetails[0].Streams) > 0 {
-				stream := jsz.AccountDetails[0].Streams[0]
-				return &stream
-			}
-			t.Error("Could not find account details")
-			return nil
-		}
-
-		checkState := func(t *testing.T) error {
-			t.Helper()
-
-			leaderSrv := c.streamLeader("js", sc.Name)
-			if leaderSrv == nil {
-				return fmt.Errorf("no leader found for stream")
-			}
-			streamLeader := getStreamDetails(t, leaderSrv)
-			var errs []error
-			for _, srv := range c.servers {
-				if srv == leaderSrv {
-					// Skip self
-					continue
-				}
-				stream := getStreamDetails(t, srv)
-				if stream == nil {
-					return fmt.Errorf("stream not found")
-				}
-
-				if stream.State.Msgs != streamLeader.State.Msgs {
-					err := fmt.Errorf("Leader %v has %d messages, Follower %v has %d messages",
-						stream.Cluster.Leader, streamLeader.State.Msgs,
-						srv, stream.State.Msgs,
-					)
-					errs = append(errs, err)
-				}
-				if stream.State.FirstSeq != streamLeader.State.FirstSeq {
-					err := fmt.Errorf("Leader %v FirstSeq is %d, Follower %v is at %d",
-						stream.Cluster.Leader, streamLeader.State.FirstSeq,
-						srv, stream.State.FirstSeq,
-					)
-					errs = append(errs, err)
-				}
-				if stream.State.LastSeq != streamLeader.State.LastSeq {
-					err := fmt.Errorf("Leader %v LastSeq is %d, Follower %v is at %d",
-						stream.Cluster.Leader, streamLeader.State.LastSeq,
-						srv, stream.State.LastSeq,
-					)
-					errs = append(errs, err)
-				}
-			}
-			if len(errs) > 0 {
-				return errors.Join(errs...)
-			}
-			return nil
-		}
-
-		checkMsgsEqual := func(t *testing.T) {
-			// These have already been checked to be the same for all streams.
-			state := getStreamDetails(t, c.streamLeader("js", sc.Name)).State
-			// Gather the stream mset from each replica.
-			var msets []*stream
-			for _, s := range c.servers {
-				acc, err := s.LookupAccount("js")
-				require_NoError(t, err)
-				mset, err := acc.lookupStream(sc.Name)
-				require_NoError(t, err)
-				msets = append(msets, mset)
-			}
-			for seq := state.FirstSeq; seq <= state.LastSeq; seq++ {
-				var expectedErr error
-				var msgId string
-				var smv StoreMsg
-				for i, mset := range msets {
-					mset.mu.RLock()
-					sm, err := mset.store.LoadMsg(seq, &smv)
-					mset.mu.RUnlock()
-					if err != nil || expectedErr != nil {
-						// If one of the msets reports an error for LoadMsg for this
-						// particular sequence, then the same error should be reported
-						// by all msets for that seq to prove consistency across replicas.
-						// If one of the msets either returns no error or doesn't return
-						// the same error, then that replica has drifted.
-						if msgId != _EMPTY_ {
-							t.Fatalf("Expected MsgId %q for seq %d, but got error: %v", msgId, seq, err)
-						} else if expectedErr == nil {
-							expectedErr = err
-						} else {
-							require_Error(t, err, expectedErr)
-						}
-						continue
-					}
-					// Only set expected msg ID if it's for the very first time.
-					if msgId == _EMPTY_ && i == 0 {
-						msgId = string(sm.hdr)
-					} else if msgId != string(sm.hdr) {
-						t.Fatalf("MsgIds do not match for seq %d: %q vs %q", seq, msgId, sm.hdr)
-					}
-				}
-			}
-		}
-
-		// Wait for test to finish before checking state.
-		wg.Wait()
-
-		// If clustered, check whether leader and followers have drifted.
-		if sc.Replicas > 1 {
-			// If we have drifted do not have to wait too long, usually it's stuck for good.
-			checkFor(t, time.Minute, time.Second, func() error {
-				return checkState(t)
-			})
-			// If we succeeded now let's check that all messages are also the same.
-			// We may have no messages but for tests that do we make sure each msg is the same
-			// across all replicas.
-			checkMsgsEqual(t)
-		}
-
-		err = checkForErr(time.Minute, time.Second, func() error {
-			var consumerPending int
-			consumers := make(map[string]int)
-			for i := 0; i < 10; i++ {
-				consumerName := fmt.Sprintf("consumer:EEEEE:%d", i)
-				ci, err := js.ConsumerInfo(sc.Name, consumerName)
-				require_NoError(t, err)
-				pending := int(ci.NumPending)
-				consumers[consumerName] = pending
-				consumerPending += pending
-			}
-
-			// Only check if there are any pending messages.
-			if consumerPending > 0 {
-				// Check state of streams and consumers.
-				si, err := js.StreamInfo(sc.Name, &nats.StreamInfoRequest{SubjectsFilter: ">"})
-				require_NoError(t, err)
-				streamPending := int(si.State.Msgs)
-				// FIXME: Num pending can be out of sync from the number of stream messages in the subject.
-				// https://github.com/nats-io/nats-server/pull/6297
-				if streamPending != consumerPending {
-					return fmt.Errorf("Unexpected number of pending messages, stream=%d, consumers=%d \n subjects: %+v\nconsumers: %+v",
-						streamPending, consumerPending, si.State.Subjects, consumers)
-				}
-			}
-			return nil
-		})
-		if err != nil {
-			t.Logf("WRN: %v", err)
-		}
-	}
-
-	// Setting up test variations below:
-	//
-	// File based with single replica and discard old policy.
-	t.Run("R1F", func(t *testing.T) {
-		params := &testParams{
-			restartAny:     true,
-			ldmRestart:     false,
-			rolloutRestart: false,
-			restarts:       1,
-		}
-		test(t, params, &nats.StreamConfig{
-			Name:        "OWQTEST_R1F",
-			Subjects:    []string{"MSGS.>"},
-			Replicas:    1,
-			MaxAge:      30 * time.Minute,
-			Duplicates:  5 * time.Minute,
-			Retention:   nats.WorkQueuePolicy,
-			Discard:     nats.DiscardOld,
-			AllowRollup: true,
-			Placement: &nats.Placement{
-				Tags: []string{"test"},
-			},
-		})
-	})
-
-	// Clustered memory based with discard new policy and max msgs limit.
-	t.Run("R3M", func(t *testing.T) {
-		params := &testParams{
-			restartAny:     true,
-			ldmRestart:     true,
-			rolloutRestart: false,
-			restarts:       1,
-			checkHealthz:   true,
-		}
-		test(t, params, &nats.StreamConfig{
-			Name:        "OWQTEST_R3M",
-			Subjects:    []string{"MSGS.>"},
-			Replicas:    3,
-			MaxAge:      30 * time.Minute,
-			MaxMsgs:     100_000,
-			Duplicates:  5 * time.Minute,
-			Retention:   nats.WorkQueuePolicy,
-			Discard:     nats.DiscardNew,
-			AllowRollup: true,
-			Storage:     nats.MemoryStorage,
-			Placement: &nats.Placement{
-				Tags: []string{"test"},
-			},
-		})
-	})
-
-	// Clustered file based with discard new policy and max msgs limit.
-	t.Run("R3F_DN", func(t *testing.T) {
-		params := &testParams{
-			restartAny:     true,
-			ldmRestart:     true,
-			rolloutRestart: false,
-			restarts:       1,
-			checkHealthz:   true,
-		}
-		test(t, params, &nats.StreamConfig{
-			Name:        "OWQTEST_R3F_DN",
-			Subjects:    []string{"MSGS.>"},
-			Replicas:    3,
-			MaxAge:      30 * time.Minute,
-			MaxMsgs:     100_000,
-			Duplicates:  5 * time.Minute,
-			Retention:   nats.WorkQueuePolicy,
-			Discard:     nats.DiscardNew,
-			AllowRollup: true,
-			Placement: &nats.Placement{
-				Tags: []string{"test"},
-			},
-		})
-	})
-
-	// Clustered file based with discard old policy and max msgs limit.
-	t.Run("R3F_DO", func(t *testing.T) {
-		params := &testParams{
-			restartAny:     true,
-			ldmRestart:     true,
-			rolloutRestart: false,
-			restarts:       1,
-			checkHealthz:   true,
-		}
-		test(t, params, &nats.StreamConfig{
-			Name:        "OWQTEST_R3F_DO",
-			Subjects:    []string{"MSGS.>"},
-			Replicas:    3,
-			MaxAge:      30 * time.Minute,
-			MaxMsgs:     100_000,
-			Duplicates:  5 * time.Minute,
-			Retention:   nats.WorkQueuePolicy,
-			Discard:     nats.DiscardOld,
-			AllowRollup: true,
-			Placement: &nats.Placement{
-				Tags: []string{"test"},
-			},
-		})
-	})
-
-	// Clustered file based with discard old policy and no limits.
-	t.Run("R3F_DO_NOLIMIT", func(t *testing.T) {
-		params := &testParams{
-			restartAny:     true,
-			ldmRestart:     true,
-			rolloutRestart: false,
-			restarts:       1,
-			checkHealthz:   true,
-		}
-		test(t, params, &nats.StreamConfig{
-			Name:       "OWQTEST_R3F_DO_NOLIMIT",
-			Subjects:   []string{"MSGS.>"},
-			Replicas:   3,
-			Duplicates: 30 * time.Second,
-			Discard:    nats.DiscardOld,
-			Placement: &nats.Placement{
-				Tags: []string{"test"},
-			},
-		})
-	})
-}
-
 func TestJetStreamClusterConsumerNRGCleanup(t *testing.T) {
 	c := createJetStreamClusterExplicit(t, "R3S", 3)
 	defer c.shutdown()
@@ -2884,289 +2240,6 @@ func TestJetStreamClusterKeyValueDesyncAfterHardKill(t *testing.T) {
 
 	err = checkState(t, c, "$G", "KV_inconsistency")
 	require_NoError(t, err)
-}
-
-func TestJetStreamClusterKeyValueSync(t *testing.T) {
-	t.Skip("Too long for CI at the moment")
-	c := createJetStreamClusterExplicit(t, "R3S", 3)
-	defer c.shutdown()
-
-	for _, s := range c.servers {
-		s.optsMu.Lock()
-		s.opts.LameDuckDuration = 15 * time.Second
-		s.opts.LameDuckGracePeriod = -15 * time.Second
-		s.optsMu.Unlock()
-	}
-	s := c.randomNonLeader()
-	connect := func(t *testing.T) (*nats.Conn, nats.JetStreamContext) {
-		return jsClientConnect(t, s)
-	}
-
-	const accountName = "$G"
-	const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-	createData := func(n int) []byte {
-		b := make([]byte, n)
-		for i := range b {
-			b[i] = letterBytes[rand.Intn(len(letterBytes))]
-		}
-		return b
-	}
-	getOrCreateKvStore := func(kvname string) (nats.KeyValue, error) {
-		_, js := connect(t)
-		kvExists := false
-		existingKvnames := js.KeyValueStoreNames()
-		for existingKvname := range existingKvnames {
-			if existingKvname == kvname {
-				kvExists = true
-				break
-			}
-		}
-		if !kvExists {
-			return js.CreateKeyValue(&nats.KeyValueConfig{
-				Bucket:   kvname,
-				Replicas: 3,
-				Storage:  nats.FileStorage,
-			})
-		} else {
-			return js.KeyValue(kvname)
-		}
-	}
-	abs := func(x int64) int64 {
-		if x < 0 {
-			return -x
-		}
-		return x
-	}
-	var counter int64
-	var errorCounter int64
-
-	checkMsgsEqual := func(t *testing.T, accountName, streamName string) error {
-		// Gather all the streams replicas and compare contents.
-		msets := make(map[*Server]*stream)
-		for _, s := range c.servers {
-			acc, err := s.LookupAccount(accountName)
-			if err != nil {
-				return err
-			}
-			mset, err := acc.lookupStream(streamName)
-			if err != nil {
-				return err
-			}
-			msets[s] = mset
-		}
-
-		str := getStreamDetails(t, c, accountName, streamName)
-		if str == nil {
-			return fmt.Errorf("could not get stream leader state")
-		}
-		state := str.State
-		for seq := state.FirstSeq; seq <= state.LastSeq; seq++ {
-			var msgId string
-			var smv StoreMsg
-			for replica, mset := range msets {
-				mset.mu.RLock()
-				sm, err := mset.store.LoadMsg(seq, &smv)
-				mset.mu.RUnlock()
-				if err != nil {
-					if err == ErrStoreMsgNotFound || err == errDeletedMsg {
-						// Skip these.
-					} else {
-						t.Logf("WRN: Error loading message (seq=%d) from stream %q on replica %q: %v", seq, streamName, replica, err)
-					}
-					continue
-				}
-				if msgId == _EMPTY_ {
-					msgId = string(sm.hdr)
-				} else if msgId != string(sm.hdr) {
-					t.Errorf("MsgIds do not match for seq %d on stream %q: %q vs %q", seq, streamName, msgId, sm.hdr)
-				}
-			}
-		}
-		return nil
-	}
-
-	keyUpdater := func(ctx context.Context, cancel context.CancelFunc, kvname string, numKeys int) {
-		kv, err := getOrCreateKvStore(kvname)
-		if err != nil {
-			t.Fatalf("[%s]:%v", kvname, err)
-		}
-		for i := 0; i < numKeys; i++ {
-			key := fmt.Sprintf("key-%d", i)
-			kv.Create(key, createData(160))
-		}
-		lastData := make(map[string][]byte)
-		revisions := make(map[string]uint64)
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			default:
-			}
-			r := rand.Intn(numKeys)
-			key := fmt.Sprintf("key-%d", r)
-
-			for i := 0; i < 5; i++ {
-				_, err := kv.Get(key)
-				if err != nil {
-					atomic.AddInt64(&errorCounter, 1)
-					if err == nats.ErrKeyNotFound {
-						t.Logf("WRN: Key not found! [%s/%s] - [%s]", kvname, key, err)
-						cancel()
-					}
-				}
-			}
-
-			k, err := kv.Get(key)
-			if err != nil {
-				atomic.AddInt64(&errorCounter, 1)
-			} else {
-				if revisions[key] != 0 && abs(int64(k.Revision())-int64(revisions[key])) < 2 {
-					lastDataVal, ok := lastData[key]
-					if ok && k.Revision() == revisions[key] && slices.Compare(lastDataVal, k.Value()) != 0 {
-						t.Logf("data loss [%s/%s][rev:%d] expected:[%v] is:[%v]", kvname, key, revisions[key], string(lastDataVal), string(k.Value()))
-					}
-				}
-				newData := createData(160)
-				revisions[key], err = kv.Update(key, newData, k.Revision())
-				if err != nil && err != nats.ErrTimeout {
-					atomic.AddInt64(&errorCounter, 1)
-				} else {
-					lastData[key] = newData
-				}
-				atomic.AddInt64(&counter, 1)
-			}
-		}
-	}
-
-	streamCount := 50
-	keysCount := 100
-	streamPrefix := "IKV"
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-	defer cancel()
-
-	// The keyUpdaters will run for less time.
-	kctx, kcancel := context.WithTimeout(context.Background(), 3*time.Minute)
-	defer kcancel()
-
-	var wg sync.WaitGroup
-	var streams []string
-	for i := 0; i < streamCount; i++ {
-		streamName := fmt.Sprintf("%s-%d", streamPrefix, i)
-		streams = append(streams, "KV_"+streamName)
-
-		wg.Add(1)
-		go func(i int) {
-			defer wg.Done()
-			keyUpdater(kctx, cancel, streamName, keysCount)
-		}(i)
-	}
-
-	debug := false
-	nc2, _ := jsClientConnect(t, s)
-	if debug {
-		go func() {
-			for range time.NewTicker(5 * time.Second).C {
-				select {
-				case <-ctx.Done():
-					return
-				default:
-				}
-				for _, str := range streams {
-					leaderSrv := c.streamLeader(accountName, str)
-					if leaderSrv == nil {
-						continue
-					}
-					streamLeader := getStreamDetails(t, c, accountName, str)
-					if streamLeader == nil {
-						continue
-					}
-					t.Logf("|------------------------------------------------------------------------------------------------------------------------|")
-					lstate := streamLeader.State
-					t.Logf("| %-10s | %-10s | msgs:%-10d | bytes:%-10d | deleted:%-10d | first:%-10d | last:%-10d |",
-						str, leaderSrv.String()+"*", lstate.Msgs, lstate.Bytes, lstate.NumDeleted, lstate.FirstSeq, lstate.LastSeq,
-					)
-					for _, srv := range c.servers {
-						if srv == leaderSrv {
-							continue
-						}
-						acc, err := srv.LookupAccount(accountName)
-						if err != nil {
-							continue
-						}
-						stream, err := acc.lookupStream(str)
-						if err != nil {
-							t.Logf("Error looking up stream %s on %s replica", str, srv)
-							continue
-						}
-						state := stream.state()
-
-						unsynced := lstate.Msgs != state.Msgs || lstate.Bytes != state.Bytes ||
-							lstate.NumDeleted != state.NumDeleted || lstate.FirstSeq != state.FirstSeq || lstate.LastSeq != state.LastSeq
-
-						var result string
-						if unsynced {
-							result = "UNSYNCED"
-						}
-						t.Logf("| %-10s | %-10s | msgs:%-10d | bytes:%-10d | deleted:%-10d | first:%-10d | last:%-10d | %s",
-							str, srv, state.Msgs, state.Bytes, state.NumDeleted, state.FirstSeq, state.LastSeq, result,
-						)
-					}
-				}
-				t.Logf("|------------------------------------------------------------------------------------------------------------------------| %v", nc2.ConnectedUrl())
-			}
-		}()
-	}
-
-	checkStreams := func(t *testing.T) {
-		for _, str := range streams {
-			checkFor(t, time.Minute, 500*time.Millisecond, func() error {
-				return checkState(t, c, accountName, str)
-			})
-			checkFor(t, time.Minute, 500*time.Millisecond, func() error {
-				return checkMsgsEqual(t, accountName, str)
-			})
-		}
-	}
-
-Loop:
-	for range time.NewTicker(30 * time.Second).C {
-		select {
-		case <-ctx.Done():
-			break Loop
-		default:
-		}
-		rollout := func(t *testing.T) {
-			for _, s := range c.servers {
-				// For graceful mode
-				s.lameDuckMode()
-				s.WaitForShutdown()
-				s = c.restartServer(s)
-
-				hctx, hcancel := context.WithTimeout(context.Background(), 15*time.Second)
-				defer hcancel()
-
-			Healthz:
-				for range time.NewTicker(2 * time.Second).C {
-					select {
-					case <-hctx.Done():
-					default:
-					}
-
-					status := s.healthz(nil)
-					if status.StatusCode == 200 {
-						break Healthz
-					}
-				}
-				c.waitOnClusterReady()
-				checkStreams(t)
-			}
-		}
-		rollout(t)
-		checkStreams(t)
-	}
-	wg.Wait()
-	checkStreams(t)
 }
 
 func TestJetStreamClusterKeyValueLastSeqMismatch(t *testing.T) {
